--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/question/dto/SurveyOptionsResponse.java
@@ -1,12 +1,0 @@
-package org.nexters.jaknaesocore.domain.question.dto;
-
-import java.util.List;
-import org.nexters.jaknaesocore.domain.question.model.SurveyOption;
-
-public record SurveyOptionsResponse(Long id, String optionContents) {
-  public static List<SurveyOptionsResponse> of(List<SurveyOption> options) {
-    return options.stream()
-        .map(option -> new SurveyOptionsResponse(option.getId(), option.getContent()))
-        .toList();
-  }
-}
