--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/question/dto/SurveyResponse.java
@@ -1,16 +1,0 @@
-package org.nexters.jaknaesocore.domain.question.dto;
-
-import java.util.List;
-import org.nexters.jaknaesocore.domain.question.model.Survey;
-
-public record SurveyResponse(
-    Long id, String contents, String surveyType, List<SurveyOptionsResponse> options) {
-
-  public static SurveyResponse of(Survey survey) {
-    return new SurveyResponse(
-        survey.getId(),
-        survey.getContent(),
-        survey.getSurveyType().name(),
-        SurveyOptionsResponse.of(survey.getOptions()));
-  }
-}
