--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/question/model/MultipleChoiceSurvey.java
@@ -1,21 +1,0 @@
-package org.nexters.jaknaesocore.domain.question.model;
-
-import jakarta.persistence.DiscriminatorValue;
-import jakarta.persistence.Entity;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@DiscriminatorValue("MULTIPLE_CHOICE")
-public class MultipleChoiceSurvey extends Survey {
-
-  public MultipleChoiceSurvey(String content, SurveyBundle surveyBundle) {
-    super(content, surveyBundle);
-  }
-
-  @Override
-  public SurveyType getSurveyType() {
-    return SurveyType.MULTIPLE_CHOICE;
-  }
-}
