--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/question/model/SurveyType.java
@@ -1,7 +1,0 @@
-package org.nexters.jaknaesocore.domain.question.model;
-
-public enum SurveyType {
-  MULTIPLE_CHOICE,
-  BALANCE,
-  ;
-}
