--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/question/repository/BundleRepository.java
@@ -1,6 +1,0 @@
-package org.nexters.jaknaesocore.domain.question.repository;
-
-import org.nexters.jaknaesocore.domain.question.model.SurveyBundle;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BundleRepository extends JpaRepository<SurveyBundle, Long> {}
