--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/question/repository/SurveyOptionRepository.java
@@ -1,6 +1,0 @@
-package org.nexters.jaknaesocore.domain.question.repository;
-
-import org.nexters.jaknaesocore.domain.question.model.SurveyOption;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SurveyOptionRepository extends JpaRepository<SurveyOption, Long> {}
