+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/survey/model/SurveyType.java
@@ -1,0 +1,7 @@
+package org.nexters.jaknaesocore.domain.survey.model;
+
+public enum SurveyType {
+  MULTIPLE_CHOICE,
+  BALANCE,
+  ;
+}
