+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/survey/repository/BundleRepository.java
@@ -1,0 +1,6 @@
+package org.nexters.jaknaesocore.domain.survey.repository;
+
+import org.nexters.jaknaesocore.domain.survey.model.SurveyBundle;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BundleRepository extends JpaRepository<SurveyBundle, Long> {}
