+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/survey/repository/SurveyOptionRepository.java
@@ -1,0 +1,6 @@
+package org.nexters.jaknaesocore.domain.survey.repository;
+
+import org.nexters.jaknaesocore.domain.survey.model.SurveyOption;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SurveyOptionRepository extends JpaRepository<SurveyOption, Long> {}
