+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/survey/repository/SurveyRepository.java
@@ -1,0 +1,6 @@
+package org.nexters.jaknaesocore.domain.survey.repository;
+
+import org.nexters.jaknaesocore.domain.survey.model.Survey;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SurveyRepository extends JpaRepository<Survey, Long> {}
