+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/survey/repository/SurveySubmissionRepository.java
@@ -1,0 +1,12 @@
+package org.nexters.jaknaesocore.domain.survey.repository;
+
+import java.util.Collection;
+import org.nexters.jaknaesocore.domain.survey.model.SurveySubmission;
+import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SurveySubmissionRepository extends JpaRepository<SurveySubmission, Long> {
+
+  @EntityGraph(attributePaths = {"survey"})
+  Collection<SurveySubmission> findByMember_IdAndDeletedAtIsNull(Long memberId);
+}
