+++ jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/survey/service/SurveyService.java
@@ -1,0 +1,32 @@
+package org.nexters.jaknaesocore.domain.survey.service;
+
+import java.util.Set;
+import lombok.RequiredArgsConstructor;
+import org.nexters.jaknaesocore.domain.survey.dto.SurveyResponse;
+import org.nexters.jaknaesocore.domain.survey.model.Survey;
+import org.nexters.jaknaesocore.domain.survey.model.SurveyBundle;
+import org.nexters.jaknaesocore.domain.survey.repository.BundleRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Service
+public class SurveyService {
+
+  private final BundleRepository bundleRepository;
+  private final SurveySubmissionReader surveySubmissionReader;
+
+  @Transactional(readOnly = true)
+  public SurveyResponse getSurvey(Long bundleId, Long memberId) {
+    SurveyBundle bundle =
+        bundleRepository
+            .findById(bundleId)
+            .orElseThrow(() -> new IllegalArgumentException("설문 번들을 찾을 수 없습니다."));
+
+    Set<Long> submittedSurveyIds = surveySubmissionReader.getSubmittedSurveyIds(memberId);
+
+    Survey unSubmittedSurvey = bundle.getUnSubmittedSurvey(submittedSurveyIds);
+
+    return SurveyResponse.of(unSubmittedSurvey);
+  }
+}
